Add files via upload
Optimized Data Retrieval and Insertion in Banking System: Developed and implemented hashing techniques to manage data in a banking system, 
addressing collision through chaining, linear probing, quadratic probing, and cubic probing.
Clustering leads to bad find/insert/delete in linear probing. So in quadratic probing we improve that.
So, quadratic probing is better as it had no Secondary Clusters.
While cubic probing further improves clustering and thus optimizing time complexity for individual search/insert/delete operations.
Worst case Time complexity of implementing the probing algorithms is 𝑂(𝑁∗𝑆),
where N is no of the keys to be inserted and S is the size of the hash table.